fix(DD-graph): Refactor graph code into helper functions
