+++ src/DataDictionary/GraphCreator.js
@@ -1,26 +1,13 @@
 import * as d3 from "d3"; 
 import React from "react";
-import { Link } from 'react-router';
-import { button } from '../theme';
-import styled from 'styled-components';
+import { color, legend_creator, add_arrows, add_links, calculate_position } from "../utils"
 
-const ToggleButton = styled.a`
-  border: 1px solid darkslategray;
-  color: darkslategray;
-  ${button};
-  position:absolute;
-  top:20px;
-  left:30px;
-  z-index: 100;
-  &:hover,
-  &:active,
-  &:focus {
-    color: black;
-    border-color: black;
-  }
-`;
-
-function create_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, svg_height_mult) {
+/**
+ * create_dd_graph: Creates a Data Dictionary graph (rectangular nodes).
+ *    Needs position as property of each node (as fraction of 1 e.g. [0.5, 0.1] 
+ *    for placement at (0.5*svg_width, 0.1*svg_height))
+ */
+function create_dd_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, svg_height_mult) {
   let max_x_pos = Math.round(1/d3.extent(nodes.map((node) => node.position[0]))[0])
   let max_y_pos = Math.round(1/d3.extent(nodes.map((node) => node.position[1]))[0])
 
@@ -51,51 +38,13 @@ function create_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, 
   let legend = svg.append("g")
     .attr("transform", "translate(" + (width-legend_width*2) + "," + padding + ")");
 
-  let defs = graph.append('svg:defs');
-  defs.append('svg:marker')
-    .attr('id', 'end-arrow')
-    .attr('viewBox', '0 -5 10 10')
-    .attr('fill', 'darkgray')
-    .attr('refX', 0)
-    .attr('refY', 0)
-    .attr('markerWidth', 6)
-    .attr('markerHeight', 6)
-    .attr('orient', 'auto')
-    .append('svg:path')
-      .attr('d', 'M0,-5L10,0L0,5');
+  let link = add_links(graph, edges)
 
-  let color = {
-    "administrative": d3.schemeCategory20[12],
-    "clinical": d3.schemeCategory20[11],
-    "biospecimen": d3.schemeCategory20[16],
-    "metadata_file": d3.schemeCategory20b[14],
-    "index_file": d3.schemeCategory20[18],
-    "notation": d3.schemeCategory20[19],
-    "data_file": d3.schemeCategory20[17],
-  }
+  add_arrows(graph)
 
-  let simulation = d3.forceSimulation()
-      .force("link", d3.forceLink().id(function(d) { return d.name; }))
-
-  let link = graph.append("g")
-    .selectAll("path")
-    .data(edges)
-    .enter().append("path")
-      .attr("stroke-width", 2)
-      .attr("marker-mid", "url(#end-arrow)")
-      .attr("stroke", "darkgray")
-      .attr("fill", "none");
-
-  // Calculate the appropriate position of each node on the graph
-  let fy_vals = []; 
-  for (let i = 0; i < nodes.length; i++) {
-    nodes[i].fx = nodes[i].position[0]*width;
-    nodes[i].fy = nodes[i].position[1]*height;
-    if (fy_vals.indexOf(nodes[i].fy) == -1) {
-      fy_vals.push(nodes[i].fy);
-    }
-  }
-  let num_rows = fy_vals.length;
+  let calc_pos_obj = calculate_position(nodes, width, height);
+  let num_rows = calc_pos_obj.fy_vals_length;
+  nodes = calc_pos_obj.nodes;
 
   let node_types = nodes.map((node) => node.name)
 
@@ -115,26 +64,6 @@ function create_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, 
           }
         }
       });
-
-  // Find all unique categories 
-  let unique_categories_array = nodes.reduce((acc, elem) => {
-    if (acc.indexOf(elem.category) === -1) {
-      acc.push(elem.category);
-    }
-    return acc;
-  }, []);
-  unique_categories_array.sort(function(a, b) {
-      a = a.toLowerCase();
-      b = b.toLowerCase();
-      if (a < b) {
-        return -1;
-      } else if (a > b) {
-        return 1;
-      } else {
-        return 0;
-      }
-    }
-  )
 
   // Add nodes to graph
   node.append("rect")
@@ -160,6 +89,9 @@ function create_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, 
       .style("font-weight", "bold")
       .text(nodes[n].name);
   }
+
+  let simulation = d3.forceSimulation()
+      .force("link", d3.forceLink().id(function(d) { return d.name; }))
 
   // Put the nodes and edges in the correct spots
   simulation
@@ -205,32 +137,19 @@ function create_graph(nodes, edges, radius=60, box_height_mult, box_width_mult, 
     }
   }
 
-  let legend_font_size = "0.9em"
-  //Make Legend
-  legend.selectAll("text")
-    .data(unique_categories_array)
-    .enter().append("text")
-    .attr("x", legend_width/2)
-    .attr("y", function(d, i) {
-      return (1.5*(2.5+i))+"em";
-    })
-    .attr("text-anchor", "middle")
-    .attr("fill", function(d) { return color[d]})
-    .style("font-size", legend_font_size)
-    .text(function(d) {return d});
-
-  legend.append("text")
-    .attr("x", legend_width/2)
-    .attr("y", 2 +"em")
-    .attr("text-anchor", "middle")
-    .text("Categories")
-    .style("font-size", legend_font_size)
-    .style("text-decoration", "underline");
+  legend_creator(legend, nodes, legend_width, color)
 }
 
+/**
+ * formatField: Recurisvely inserts newline characters into strings that are 
+ *    too long after underscores
+ */
 function formatField(name) {
   if (name.length > 20) {
     let split_name = name.split('_')
+    if (split_name.length == 1) {
+      return name
+    }
     let mid = Math.ceil(split_name.length/2)
     let begin = split_name.slice(0, mid).join('_')
     let end = split_name.slice(mid).join('_')
@@ -246,6 +165,10 @@ function formatField(name) {
   }
 }
 
+/** 
+ * formatType: Turn different ways used to represent type in data dictionary 
+ *    into a string
+ */
 function formatType(type) {
   if (typeof type == 'string') {
     return type
@@ -276,6 +199,11 @@ function formatType(type) {
   }
 }
 
+/**
+ * add_tables: Add tables to data dictionary graph.
+ *    Also hides the node names rendered by svg and replaces them with non-svg
+ *    text so they remain clickable
+ */
 function add_tables(nodes, box_width, box_height, svg_width, svg_height) {
   let table_div = d3.select("#graph_wrapper")
   .append('div')
@@ -354,7 +282,7 @@ function add_tables(nodes, box_width, box_height, svg_width, svg_height) {
   }
 }
 
-function create_full_graph(nodes, edges) {
+export function create_full_graph(nodes, edges) {
   let radius = 60
   let box_height = radius * 4
   let box_width = radius * 4
@@ -365,7 +293,7 @@ function create_full_graph(nodes, edges) {
   let svg_width = max_x_pos * radius * 5
   let svg_height = max_y_pos * radius * 5
   
-  create_graph(nodes, edges, radius, 4, 4, 5)
+  create_dd_graph(nodes, edges, radius, 4, 4, 5)
 
   if (document.getElementById("table_wrapper") != null) {
     document.getElementById("table_wrapper").remove();
@@ -374,103 +302,6 @@ function create_full_graph(nodes, edges) {
   add_tables(nodes, box_width, box_height, svg_width, svg_height)
 };
 
-function create_abridged_graph(nodes, edges) {
-  create_graph(nodes, edges, 60, 1.5, 3, 3)
+export function create_abridged_graph(nodes, edges) {
+  create_dd_graph(nodes, edges, 60, 1.5, 3, 3)
 };
-
-export default class CreateGraph extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      full_toggle: true
-    }
-    this.handleClick = this.handleClick.bind(this);
-  }
-  componentDidMount() {
-    if (this.state.full_toggle) {
-      create_full_graph(this.props.nodes, this.props.edges)
-    } else {
-      document.getElementById("table_wrapper").remove()
-      create_abridged_graph(this.props.nodes, this.props.edges)
-    }
-  }
-  componentDidUpdate() {
-    if (this.state.full_toggle) {
-      create_full_graph(this.props.nodes, this.props.edges)
-    } else {
-      document.getElementById("table_wrapper").remove()
-      create_abridged_graph(this.props.nodes, this.props.edges)
-    }
-  }
-  handleClick() {
-    this.setState(prevState => ({
-      full_toggle: !prevState.full_toggle,
-    }));
-  }
-  render() {
-    let nodes = this.props.nodes
-    let edges = this.props.edges
-
-    let root = "program"
-    let queue = [];
-    let layout = [];
-    let placed = [];
-    let layout_level = 0;
-
-    queue.push(root);
-    layout.push([root]);
-    while(queue.length != 0) {
-      let query = queue.shift(); //breadth first
-      for (let i = 0; i < edges.length; i++) {
-        if (edges[i].target == query || edges[i].target.name == query) {
-          if (!layout[layout_level+1]) {
-            layout[layout_level+1] = [];
-          } 
-          queue.push(edges[i].source);
-          if ((layout[layout_level+1].indexOf(edges[i].source) == -1) && (placed.indexOf(edges[i].source) == -1)) {
-            if (layout[layout_level+1].length >= 3) {
-              layout_level += 1;
-              if (!layout[layout_level+1]) {
-                layout[layout_level+1] = [];
-              }
-            }
-            layout[layout_level+1].push(edges[i].source);
-            placed.push(edges[i].source);
-          }
-        }
-      }
-      placed.push(query);
-    }
-
-    for (let i = 0; i < layout.length; i++) {
-      for (let j = 0; j < layout[i].length; j++) {
-        for (let k = 0; k < nodes.length; k++) {
-          if (nodes[k].name == layout[i][j]) {
-            nodes[k].position = [(j+1)/(layout[i].length+1),(i+1)/(layout.length+1)];
-            nodes[k].position_index = [j,i];
-            break;
-          }
-        }
-      }
-    }
-
-    const divStyle = {
-      width: "inherit",
-      backgroundColor: "#f4f4f4",
-      margin: "0 auto",
-      textAlign: "center",
-      position: "relative"
-    }
-    return (
-      <div>
-        <Link to={'/dd'}> Explore dictionary as a table </Link>
-        <p style={{"fontSize": "75%", "marginTop": "1em"}}> <span style={{"fontWeight": "bold", "fontStyle": "italic"}}> Bold, italicized</span> properties are required</p>
-        <div style={divStyle} id="graph_wrapper">
-          <ToggleButton onClick={this.handleClick}>Toggle view</ToggleButton>
-          <svg id="data_model_graph">
-          </svg>
-        </div>
-      </div>
-    );
-  }
-}
